requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
groq>=0.4.0
aiohttp>=3.9.0
pymongo>=4.6.0
